--- game_files/ahoy/Assets/LookAtCam.cs
@@ -1,14 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class LookAtCam : MonoBehaviour
-{
-    public Transform boat;
-
-    // Update is called once per frame
-    void Update()
-    {
-        transform.LookAt(boat.position);
-    }
-}
