+++ game_files/ahoy/Assets/Materials/NewLayer.terrainlayer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fee1bc88aa193401db6bf194d9e5c277
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 8574412962073106934
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
