+++ game_files/ahoy/Assets/Materials/Rock.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0fe2286435a6d46e999a88317be037bd
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
