+++ game_files/ahoy/Assets/Materials/Skybox.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b9ba6b72edfbc48a8b1c70bf7104c0a6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
