+++ game_files/ahoy/Assets/Materials/Textures/NewLayer 1.terrainlayer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8c07da4c87e6749c58043da6b70fe7de
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 8574412962073106934
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
