+++ game_files/ahoy/Assets/Materials/Textures/NewLayer 2.terrainlayer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b5d2752e6fd5644b38ab26cb9d02a91b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 8574412962073106934
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
