+++ game_files/ahoy/Assets/New Terrain.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1633e474d667c4ce99ebfa2d8b488a6d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 15600000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
