+++ game_files/ahoy/Assets/Physics Materials/Ground.physicMaterial
@@ -1,0 +1,14 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!134 &13400000
+PhysicMaterial:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: Ground
+  dynamicFriction: 0.3
+  staticFriction: 0.3
+  bounciness: 0
+  frictionCombine: 0
+  bounceCombine: 0
