+++ game_files/ahoy/Assets/Scripts/Camera/FollowCamera.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d7b14e80997db465fb0c8b7bf495da73
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
