+++ game_files/ahoy/Assets/Scripts/Camera/LookAtCam.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8b64087b303ec43ef8491561433b6e2d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
