--- game_files/ahoy/Assets/Scripts/DebugCam.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 589d7e602a2094010ba0ebb4d7757e78
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
